PD-6967 updated tests to pass linting rules as well as made tests unit tests
