+++ lib/control/validation/signature.js
@@ -2,23 +2,27 @@
 
 const FORBIDDEN = 403;
 
-const server = require('../v1/signature');
-
-exports.isValidsignature = function(req, res, next) {
-
-  signatureServer.validate({
+exports.signatureValidate = function (req, res, next, server) {
+  server.validate({
     signature: req.body.signature,
     data: req.body.document
   },
   function(err, status) {
+    console.log('error is: ')
+    console.log(err)
+    console.log('\n\n\n')
+    console.log('status is: ');
+    console.log(status);
+    console.log('\n\n\n')
+    console.log('valid is: ');
+    console.log(status.valid);
     if (err) { return next(err) }
     if (!status.valid) {
       return res.status(FORBIDDEN).json({
         code: FORBIDDEN,
-        status: status,
         error: 'Signature is not valid'
       });
     }
     return next();
-  });
-};
+  })
+}
