--- test/lib/helpers.js
@@ -1,7 +1,0 @@
-'use strict';
-
-/* eslint-env mocha */
-/* global Config, Log */
-/* eslint-disable rapid7/static-magic-numbers, max-nested-callbacks */
-global.Log = new (require('winston').Logger);
-global.Config = require('nconf');
